requirements.txt
Django==2.0.7
djangorestframework==3.8.2
djangorestframework-simplejwt==3.2.3